refactor: move to validator
