+++ api/src/controllers/program.js
@@ -5,13 +5,13 @@ const { capture } = require("../sentry");
 
 const ProgramObject = require("../models/program");
 const { ERRORS } = require("../utils");
-const { validateId, validateString } = require("../utils/defaultValidate");
-const validateFromReferent = require("../utils/referent");
-const validateFromYoung = require("../utils/young");
+const { validateId, validateString } = require("../utils/defaultValidator");
+const referentValidator = require("../utils/validator/referent");
+const youngValidator = require("../utils/validator/young");
 
 router.post("/", async (req, res) => {
   try {
-    const { error, value: checkedProgram } = validateFromYoung.validateProgram(req.body);
+    const { error, value: checkedProgram } = youngValidator.validateProgram(req.body);
     if (error) return res.status(400).send({ ok: false, code: ERRORS.INVALID_BODY, error });
     const data = await ProgramObject.create(checkedProgram);
     return res.status(200).send({ ok: true, data });
@@ -23,7 +23,7 @@ router.post("/", async (req, res) => {
 
 router.put("/", passport.authenticate("referent", { session: false }), async (req, res) => {
   try {
-    const { error: errorProgram, value: checkedProgram } = validateFromReferent.validateProgram(req.body);
+    const { error: errorProgram, value: checkedProgram } = referentValidator.validateProgram(req.body);
     const { error: errorId, value: checkedId } = validateId(req.body._id);
     if (errorProgram || errorId) return res.status(400).send({ ok: false, code: ERRORS.INVALID_BODY, error });
     let obj = checkedProgram;
